abstract out clearAllCells
